Set global tabstop to 8
Ideally it would be great to never have tabs, but sometimes they may
occur and in that case it would be better to be aware of their presence.

By setting tabstop to 8 (the default value), vim will always display
8 spaces when it encounters tabs in the code, so they can be spotted
easily (and hopefully removed).

Having expandtab set will ensure that vim will never insert tabs in the
code, so the code will be always properly formatted to the desired
shiftwidth value.

However, sometimes you can't remove the tabs (you may not own the code
so different conventions may be used), in that case, you may want to
disable expandtab and set tabstop to a different value to make sure to
use tabs properly as in the (bloated) code convention.